requirements_py3-12-4.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.6
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.1.0
narwhals==1.18.4
numpy==2.2.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.1
pyarrow==18.1.0
pydantic==2.10.3
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.33.0
watchdog==6.0.0
watchfiles==1.0.3
websockets==14.1
